Refactored serializers
