+++ app/repositories/managers/__init__.py
@@ -1,8 +1,8 @@
-from .base_manager import BaseManager
-from .beverage_manager import BeverageManager
-from .index_manager import IndexManager
-from .ingredient_manager import IngredientManager
-from .mock_manager import MockManager
-from .order_manager import OrderManager
-from .report_manager import ReportManager
-from .size_manager import SizeManager
+from .base import BaseManager
+from .beverage import BeverageManager
+from .index import IndexManager
+from .ingredient import IngredientManager
+from .mock import MockManager
+from .order import OrderManager
+from .report import ReportManager
+from .size import SizeManager
