+++ app/repositories/managers/base.py
