+++ app/repositories/managers/beverage.py
@@ -1,6 +1,6 @@
 from typing import Sequence
 
-from .base_manager import BaseManager
+from .base import BaseManager
 from ..models import Beverage
 from ..serializers import BeverageSerializer
