+++ app/repositories/managers/size.py
@@ -1,4 +1,4 @@
-from .base_manager import BaseManager
+from .base import BaseManager
 from ..models import Size
 from ..serializers import SizeSerializer
