--- app/repositories/serializers.py
@@ -1,73 +1,0 @@
-from app.plugins import ma
-from .models import Ingredient, Beverage, Size, Order, IngredientOrderDetail, BeverageOrderDetail
-
-
-class IngredientSerializer(ma.SQLAlchemyAutoSchema):
-
-    class Meta:
-        model = Ingredient
-        load_instance = True
-        fields = ('_id', 'name', 'price')
-
-
-class BeverageSerializer(ma.SQLAlchemyAutoSchema):
-
-    class Meta:
-        model = Beverage
-        load_instance = True
-        fields = ('_id', 'name', 'price')
-
-
-class SizeSerializer(ma.SQLAlchemyAutoSchema):
-
-    class Meta:
-        model = Size
-        load_instance = True
-        fields = ('_id', 'name', 'price')
-
-
-class IngredientOrderDetailSerializer(ma.SQLAlchemyAutoSchema):
-
-    ingredient = ma.Nested(IngredientSerializer)
-
-    class Meta:
-        model = IngredientOrderDetail
-        load_instance = True
-        fields = (
-            'ingredient_price',
-            'ingredient'
-        )
-
-
-class BeverageOrderDetailSerializer(ma.SQLAlchemyAutoSchema):
-
-    beverage = ma.Nested(BeverageSerializer)
-
-    class Meta:
-        model = BeverageOrderDetail
-        load_instance = True
-        fields = (
-            'beverage_price',
-            'beverage'
-        )
-
-class OrderSerializer(ma.SQLAlchemyAutoSchema):
-    size = ma.Nested(SizeSerializer)
-    ingredient_detail = ma.Nested(IngredientOrderDetailSerializer, many=True)
-    beverage_detail = ma.Nested(BeverageOrderDetailSerializer, many=True)
-
-    class Meta:
-        model = Order
-        load_instance = True
-        fields = (
-            '_id',
-            'client_name',
-            'client_dni',
-            'client_address',
-            'client_phone',
-            'date',
-            'total_price',
-            'size',
-            'ingredient_detail',
-            'beverage_detail'
-        )
